+++ apps/taxes/src/modules/taxjar/taxjar-calculate-taxes/taxjar-calculate-taxes-adapter.test.ts
@@ -1,0 +1,5 @@
+import { describe } from "vitest";
+
+describe.todo("TaxJarResponseTransformer", () => {});
+
+describe.todo("TaxJarCalculateTaxesAdapter", () => {});
